Fix global component loading in production builds.
Instead of maintaining an index.js which re-exports all the
vue.js modules in the global-components directory, use
webpack's `require.context` loader to iterate over all
the contained files and register the components from there.

The previous method constructed the component names from
the __file property of the module objects, but this is only
available in debug builds and so the site failed to load
when deployed.