Restore explicit loading of global components.
Patrick Altman prefers an explicit export list rather than
loading everything in the directory.

Restore global-components/index.js with an explicit list of
default exports, which was removed in the previous commit,
and revert to Jacob Wegner's loading solution from commit
16f106527bef04893ecfbd94b64808f5fdd323e1.

Instead of deriving the component name from the debug-only
__file property, export the desired components as properties
of the default object and use Object.entries to enumerate over
them, using the property name, and hence module import name, as the
registered component name.